ui/backend/requirements.txt
# Backend dependencies for Podcast Knowledge UI
# Minimal requirements for resource-constrained environments

# FastAPI framework and server
fastapi==0.115.12  # Latest stable version from context7
uvicorn[standard]==0.34.0  # ASGI server with standard extras

# Database connectivity
neo4j==5.28.0  # Neo4j Python driver
neo4j-graphrag==1.7.0  # Neo4j GraphRAG for RAG functionality

# Configuration reading
pyyaml==6.0.2  # For reading podcast configuration

# RAG dependencies
google-generativeai==0.3.2  # For embeddings via Gemini (same version as seeding pipeline)
groq==0.11.0  # For Groq LLM responses